backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
python-dotenv==1.0.1

# Phase 2 — scraper
playwright==1.49.0
playwright-stealth==2.0.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Phase 3 — reporting
pandas==2.2.3
openpyxl==3.1.5
jinja2==3.1.4

# Phase 4 — AI
google-genai==1.12.1
